Remove API versioning logic
We've not yet had projects at abtion with multiple API versions.
And if we need versioned APIs it's not a big feat to add it later.
