+++ app/controllers/api/users_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Api
+  class UsersController < Api::BaseController
+    before_action :require_authentication!
+
+    def show
+      render json: UserSerializer.new(current_user).serializable_hash
+    end
+  end
+end
