--- lib/constraints/api_version.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Constraints
-  class ApiVersion
-    attr_reader :version
-
-    def initialize(version)
-      @version = version
-    end
-
-    def matches?(request)
-      return false unless request&.headers
-
-      request.headers["X-Api-Version"] == version
-    end
-  end
-end
